+++ CHANGELOG.md
@@ -70,6 +70,7 @@ To learn more about active deprecations, we recommend checking [GitHub Discussio
 ### Improvements
 
 - **General**: Internal gRPC connection's certificates are hot reloaded ([#6756](https://github.com/kedacore/keda/pull/6756))
+- **Kafka Scaler**: New `useMessageCountOnInvalidOffset` param to add ability to use message count in the topic as a metric (enabled when consumer offset is not valid and offsetResetPolicy is set to earliest) ([#6775](https://github.com/kedacore/keda/pull/6775))
 - **Metrics API**: Support multiple auth methods simultaneously in Metrics API scaler ([#6642](https://github.com/kedacore/keda/issues/6642))
 
 ### Fixes